requirements.txt
streamlit==1.28.2
pandas==2.1.3
mysql-connector-python==8.2.0 